# trader_interactive/public/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "L5 Swagger API",
        "description": "L5 Swagger API description",
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "schemes": [
        "http"
    ],
    "paths": {
        "/api/listings": {
            "get": {
                "tags": [
                    "listings"
                ],
                "summary": "Get list of listings",
                "description": "Returns list of listings",
                "operationId": "getListings",
                "responses": {
                    "200": {
                        "description": "Listing Details Found"
                    },
                    "500": {
                        "description": "Listing Details Not Found"
                    }
                }
            }
        },
        "/api/listing-details/id={id}": {
            "get": {
                "tags": [
                    "listing-details"
                ],
                "summary": "Get Individual Listing Details",
                "description": "Returns Individual Listing Details",
                "operationId": "getListingDetails",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "listing id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Individual Listing Details Found"
                    },
                    "500": {
                        "description": "Individual Listing Details Not Found"
                    }
                }
            }
        },
        "/api/search-listing-details/make={make}/model={model}/min_price={min_price}/max_price={max_price}": {
            "get": {
                "tags": [
                    "search-listing-details"
                ],
                "summary": "Get Searched Listing Details",
                "description": "Returns Searched Listing Details",
                "operationId": "getUserListings",
                "parameters": [
                    {
                        "name": "make",
                        "in": "path",
                        "description": "make",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "model",
                        "in": "path",
                        "description": "model name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "min_price",
                        "in": "path",
                        "description": "minimum price",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "max_price",
                        "in": "path",
                        "description": "maximum price",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Searched Listing Details Found"
                    },
                    "500": {
                        "description": "Searched Listing Details Not Found"
                    }
                }
            }
        },
...
        "/api/contact": {
            "post": {
                "tags": [
                    "contact"
                ],
                "summary": "Sending Email",
                "description": "Sending Email",
                "operationId": "contact",
                "parameters": [
                    {
                        "name": "msg",
                        "in": "body",
                        "description": "msg",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "example": {
                                "msg": "john"
                            }
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message Send Successfully"
                    }
                }
            }
        },
        "/api/reviews/id={id}": {
            "get": {
                "tags": [
                    "reviews"
                ],
                "summary": "Get Individual Listing Review Details",
                "description": "Returns Individual Listing Review Details",
                "operationId": "getUserReviewDetails",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "seller id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Individual Listing Review Details Found"
                    },
                    "500": {
                        "description": "Individual Listing Review Details Not Found"
                    }
                }
            }
        },
...
        "/api/seller-details/id={id}": {
            "get": {
                "tags": [
                    "seller-details"
                ],
                "summary": "Get Individual Seller Details",
                "description": "Returns Individual Seller Details",
                "operationId": "getSellerDetails",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "seller id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Individual Seller Details Found"
                    },
                    "500": {
                        "description": "Individual Seller Details Not Found"
                    }
                }
            }
        },
...
        "/api/users": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Get list of users",
                "description": "Returns list of users",
                "operationId": "getUsers",
                "responses": {
                    "200": {
                        "description": "User Details Found"
                    },
                    "500": {
                        "description": "User Details Not Found"
                    }
                }
            }
        },
...
    },
...
}